feat: process multi-theme for building token sets (WIP)
